Migrate delete function.
