--- src/main/java/com/sebastian/ems/controller/EmployeeController.java
@@ -1,55 +1,0 @@
-package com.sebastian.ems.controller;
-
-import com.sebastian.ems.model.Employee;
-import com.sebastian.ems.service.EmployeeService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@Controller
-public class EmployeeController {
-
-    @Autowired
-    private EmployeeService employeeService;
-
-    // display list of employees
-    @GetMapping("/")
-    public String viewHomePage(Model model) {
-        return findPaginated(1, "firstName", "asc", model);
-    }
-
-    @GetMapping("/deleteEmployee/{id}")
-    public String deleteEmployee(@PathVariable (value = "id") long id) {
-
-        // call delete employee method
-        this.employeeService.deleteEmployeeById(id);
-        return "redirect:/";
-    }
-
-
-    @GetMapping("/page/{pageNo}")
-    public String findPaginated(@PathVariable (value = "pageNo") int pageNo,
-                                @RequestParam("sortField") String sortField,
-                                @RequestParam("sortDir") String sortDir,
-                                Model model) {
-        int pageSize = 5;
-
-        Page<Employee> page = employeeService.findPaginated(pageNo, pageSize, sortField, sortDir);
-        List<Employee> listEmployees = page.getContent();
-
-        model.addAttribute("currentPage", pageNo);
-        model.addAttribute("totalPages", page.getTotalPages());
-        model.addAttribute("totalItems", page.getTotalElements());
-
-        model.addAttribute("sortField", sortField);
-        model.addAttribute("sortDir", sortDir);
-        model.addAttribute("reverseSortDir", sortDir.equals("asc") ? "desc" : "asc");
-
-        model.addAttribute("listEmployees", listEmployees);
-        return "index";
-    }
-}
