--- src/main/java/com/sebastian/ems/repository/EmployeeRepository.java
@@ -1,10 +1,0 @@
-package com.sebastian.ems.repository;
-
-import com.sebastian.ems.model.Employee;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-
-}
