--- src/main/java/com/sebastian/ems/service/EmployeeService.java
@@ -1,12 +1,0 @@
-package com.sebastian.ems.service;
-
-import com.sebastian.ems.model.Employee;
-import org.springframework.data.domain.Page;
-
-import java.util.List;
-
-public interface EmployeeService {
-    List<Employee> getAllEmployees();
-    void deleteEmployeeById(long id);
-    Page<Employee> findPaginated(int pageNo, int pageSize, String sortField, String sortDirection);
-}
