--- src/main/java/com/sebastian/ems/service/EmployeeServiceImpl.java
@@ -1,39 +1,0 @@
-package com.sebastian.ems.service;
-
-import com.sebastian.ems.model.Employee;
-import com.sebastian.ems.repository.EmployeeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class EmployeeServiceImpl implements EmployeeService {
-
-    @Autowired
-    private EmployeeRepository employeeRepository;
-
-    @Override
-    public List<Employee> getAllEmployees() {
-        return employeeRepository.findAll();
-    }
-
-    @Override
-    public void deleteEmployeeById(long id) {
-        this.employeeRepository.deleteById(id);
-    }
-
-    @Override
-    public Page<Employee> findPaginated(int pageNo, int pageSize, String sortField, String sortDirection) {
-        Sort sort = sortDirection.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortField).ascending() :
-                Sort.by(sortField).descending();
-
-        Pageable pageable = PageRequest.of(pageNo - 1, pageSize, sort);
-        return this.employeeRepository.findAll(pageable);
-    }
-}
